bugfix: fixed typing and annotation issue using ClassVar:
The error is originating from the use of pydantic when defining the model. The issue is with the attribute core_chips which is expected to have a type annotation. The error suggests that core_chips was found without an annotation, but it seems like it's trying to use typing.List[datamodel.models.core_chip.CoreChip].
Specify ignored types, you'll typically do that in a Config inner class within your Pydantic model. This allows you to customize certain behaviors of the model.